Refactored monkey patch directory layout
Vagrant monkey patches folder structure now matches that of vagrant 1.2 release. Also removed global helper require form plugin.rb.
