+++ lib/vagrant-windows/guest/cap/mount_virtualbox_shared_folder.rb
@@ -1,3 +1,5 @@
+require_relative '../../helper'
+
 module VagrantWindows
   module Guest
     module Cap
