+++ lib/vagrant-windows/monkey_patches/plugins/providers/virtualbox/driver/version_4_2.rb
