+++ lib/vagrant-windows/monkey_patches/plugins/provisioners/chef/provisioner/chef_client.rb
@@ -1,0 +1,1 @@
+# TODO: support Chef client
