+++ lib/vagrant-windows/monkey_patches/plugins/provisioners/puppet/provisioner/puppet.rb
