+++ lib/vagrant-windows/monkey_patches/plugins/provisioners/puppet/provisioner/puppet_server.rb
@@ -1,0 +1,1 @@
+#TODO: Support Puppet server
