+++ lib/vagrant-windows/monkey_patches/plugins/provisioners/shell/provisioner.rb
@@ -1,4 +1,5 @@
 require "#{Vagrant::source_root}/plugins/provisioners/shell/provisioner"
+require_relative '../../../../helper'
 
 module VagrantPlugins
   module Shell
