+++ lib/vagrant-windows/plugin.rb
@@ -10,27 +10,27 @@ if Vagrant::VERSION < "1.1.0"
   raise "The Vagrant Windows plugin is only compatible with Vagrant 1.1+"
 end
 
-require_relative 'helper'
-
 if Vagrant::VERSION >= "1.2.0"
   # Monkey Patch the virtualbox share_folders action to make valid share names on windows
   require_relative "monkey_patches/plugins/providers/virtualbox/action/share_folders"
 end
 
-# Monkey patch the vbox42 driver 
-require_relative "monkey_patches/vbox_42_driver"
+# Monkey patch the vbox42 driver to support read mac addresses
+require_relative "monkey_patches/plugins/providers/virtualbox/driver/version_4_2"
 
 # Monkey Patch the VM object to support multiple channels, i.e. WinRM
-require_relative "monkey_patches/machine"
+require_relative "monkey_patches/lib/vagrant/machine"
 
-# Monkey patch the Puppet provisioner to support PowerShell/Windows
-require_relative "monkey_patches/puppet"
+# Monkey patch the Puppet provisioners to support PowerShell/Windows
+require_relative "monkey_patches/plugins/provisioners/puppet/provisioner/puppet"
+require_relative "monkey_patches/plugins/provisioners/puppet/provisioner/puppet_server"
 
-# Monkey patch the Chef-Solo provisioner to support PowerShell/Windows
-require_relative "monkey_patches/chef_solo"
+# Monkey patch the Chef provisioners to support PowerShell/Windows
+require_relative "monkey_patches/plugins/provisioners/chef/provisioner/chef_solo"
+require_relative "monkey_patches/plugins/provisioners/chef/provisioner/chef_client"
 
 # Monkey patch the shell provisioner to support PowerShell/batch/exe/Windows/etc
-require_relative "monkey_patches/provisioner"
+require_relative "monkey_patches/plugins/provisioners/shell/provisioner"
 
 
 module VagrantWindows
